+++ hugr-model/src/v0/binary/read.rs
@@ -185,6 +185,30 @@ fn read_operation<'a>(
             });
             model::Operation::DeclareAlias { decl }
         }
+        Which::ConstructorDecl(reader) => {
+            let reader = reader?;
+            let name = bump.alloc_str(reader.get_name()?.to_str()?);
+            let params = read_list!(bump, reader, get_params, read_param);
+            let r#type = model::TermId(reader.get_type());
+            let decl = bump.alloc(model::ConstructorDecl {
+                name,
+                params,
+                r#type,
+            });
+            model::Operation::DeclareConstructor { decl }
+        }
+        Which::OperationDecl(reader) => {
+            let reader = reader?;
+            let name = bump.alloc_str(reader.get_name()?.to_str()?);
+            let params = read_list!(bump, reader, get_params, read_param);
+            let r#type = model::TermId(reader.get_type());
+            let decl = bump.alloc(model::OperationDecl {
+                name,
+                params,
+                r#type,
+            });
+            model::Operation::DeclareOperation { decl }
+        }
         Which::Custom(name) => model::Operation::Custom {
             operation: read_global_ref(bump, name?)?,
         },
@@ -210,6 +234,7 @@ fn read_region<'a>(
     let kind = match reader.get_kind()? {
         hugr_capnp::RegionKind::DataFlow => model::RegionKind::DataFlow,
         hugr_capnp::RegionKind::ControlFlow => model::RegionKind::ControlFlow,
+        hugr_capnp::RegionKind::Module => model::RegionKind::Module,
     };
 
     let sources = read_list!(bump, reader, get_sources, read_link_ref);